final commit

+++ examples/udp_rx_tx/udp.c
@@ -250,5 +250,5 @@ void print_prr(uint32_t pkt_rcv, uint32_t num_pkts)
      * of packets you actually received. Calculate the Packet Reception Ratio 
      * and print it out 
      */
-    printf("Ratio: %lf\n",pkt_rcv/num_pkts); 
+    printf("Ratio: %f\n",(float)pkt_rcv/(float)num_pkts); 
 }
